Move version data types out of daml-project-config
