+++ daml-assistant/BUILD.bazel
@@ -24,7 +24,9 @@ da_haskell_library(
     ],
     src_strip_prefix = "daml-project-config",
     visibility = ["//visibility:public"],
-    deps = ["//sdk-version/hs:sdk-version-class-lib"],
+    deps = [
+        "//libs-haskell/da-version-types",
+    ],
 )
 
 da_haskell_library(
