+++ libs-haskell/da-version-types/BUILD.bazel
@@ -1,0 +1,26 @@
+# Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load("//bazel_tools:haskell.bzl", "da_haskell_library", "da_haskell_test", "generate_and_track_cabal")
+load("@os_info//:os_info.bzl", "is_windows")
+
+da_haskell_library(
+    name = "da-version-types",
+    srcs = glob(["src/**/*.hs"]),
+    hackage_deps = [
+        "base",
+        "extra",
+        "lens",
+        "safe-exceptions",
+        "semver",
+        "text",
+        "yaml",
+    ],
+    src_strip_prefix = "src",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//sdk-version/hs:sdk-version-class-lib",
+    ],
+)
+
+generate_and_track_cabal("da-version-types")
