+++ libs-haskell/da-version-types/da-version-types.cabal
@@ -1,9 +1,9 @@
 cabal-version: 2.4
 -- This file was autogenerated
-name: daml-project-config
+name: da-version-types
 build-type: Simple
 version: 0.1.15.0
-synopsis: daml-project-config
+synopsis: da-version-types
 license: Apache-2.0
 author: Digital Asset
 maintainer: Digital Asset
@@ -17,24 +17,18 @@ source-repository head
 
 library
     default-language: Haskell2010
-    hs-source-dirs: daml-project-config
+    hs-source-dirs: src
     build-depends:
-      -- @stackage//:aeson
-      aeson,
       -- @stackage//:base
       base,
-      -- //libs-haskell/da-version-types:da-version-types
-      da-version-types,
-      -- @stackage//:directory
-      directory,
       -- @stackage//:extra
       extra,
-      -- @stackage//:filepath
-      filepath,
       -- @stackage//:lens
       lens,
       -- @stackage//:safe-exceptions
       safe-exceptions,
+      -- //sdk-version/hs:sdk-version-class-lib
+      sdk-version-class-lib,
       -- @stackage//:semver
       semver,
       -- @stackage//:text
@@ -42,10 +36,7 @@ library
       -- @stackage//:yaml
       yaml,
     exposed-modules:
-      DA.Daml.Project.Config
-      DA.Daml.Project.Consts
-      DA.Daml.Project.Types
-      DA.Daml.Project.Util
+      DA.Daml.Version.Types
     default-extensions:
       BangPatterns
       DeriveDataTypeable
